poll refactoring
