+++ app/backend/src/poll/poll.module.ts
@@ -2,12 +2,13 @@ import { Module } from '@nestjs/common';
 import { PollService } from './poll.service';
 import { PollController } from './poll.controller';
 import { Poll } from './entities/poll.entity';
+import { Option } from '../option/entities/option.entity'; // Import Option entity
+import { Tag } from '../tag/entities/tag.entity'; // Import Tag entity
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Option } from './entities/option.entity';
-import { Tag } from './entities/tag.entity';
+
 
 @Module({
-  imports: [TypeOrmModule.forFeature([Poll, Tag, Option])],
+  imports: [TypeOrmModule.forFeature([Poll, Option, Tag])],
   controllers: [PollController],
   providers: [PollService],
   exports: [PollService],
